Reconsider reexports
The reexports of MonadBase(ControlIdentity) are mainly so that you can
use the class signatures in your modules without having to also
specify transitive dependencies.

So re-exporting them once from this library is enough.

We keep the reexport of websocket types in the `Websocket.Trans`
module for the same reason.

For the `Wai` types we can expect that somebody depending on
`wai-control` will also depend on `wai` already, so that’s not as much
of a problem.